refactor - move Create Analysis Dialog to separate component
